Added reading multiple games from PGN and arguments to change the direction of analysis
If a PGN file has multiple games inside the script reads and analyses all of them. 
If the script is used with '--backwards' or '-b' the analysis is done from the last move of the game back to the first. This makes information from later positions available in the hashtable, and also can affect the best move selection from the engine.